Refactor renamer into its own module
